fix: Avoid `wrong-type-argument` error on `read-char`.
The `read-char` used by the `avy-goto-*` commands can return Emacs
events which will not pass `characterp` (i.e. `A-v` will generate an
event `4194422`).  As the `avy-goto-*` commands are generally
converting the value returned by `(let ((char (read-char ...))))` to a
string (via `(string char)`), this causes a `wrong-type-argument`
error to be thrown from these functions.  Wrapping the `read-char`
with `event-basic-type` avoids this error.